YOLO from scratch/requirements.txt
torch
torchvision
pandas
matplotlib
seaborn
tqdm
xml
opencv-python
numpy
Pillow
